Working version w/ cleanups
