+++ examples/morphir-elm-projects/unit-test-framework/example-project-tests/ReadMe.md
@@ -1,3 +1,3 @@
 ## Examples
 
-This project exists to let use test the unit testing functionality itself, and as a reference for how tests can be written.
+This project exists to let us test the unit testing functionality itself, and as a reference for how tests can be written.
