+++ examples/morphir-elm-projects/unit-test-framework/example-project-tests/morphir.json
@@ -2,7 +2,7 @@
     "name": "ExampleTests",
     "sourceDirectory": "src",
     "localDependencies": [
-        "../../../../elm-sdks/morphir-unit-test/morphir-ir.json",
+        "../../../../morphir-elm/sdks/morphir-unit-test/morphir-ir.json",
         "../example-project/morphir-ir.json"
     ],
     "exposedModules": [
