+++ morphir-elm/sdks/morphir-unit-test/morphir-hashes.json
@@ -1,4 +1,4 @@
 {
-    "src/Morphir/UnitTest/Expect.elm": "a740a65d8fd6cf0c11fd9ae9e156ff5b",
-    "src/Morphir/UnitTest/Test.elm": "53584b6062eb5505f6585c25054d289e"
+    "src/Morphir/UnitTest/Expect.elm": "ab16273790158c39cc1d57201476f924",
+    "src/Morphir/UnitTest/Test.elm": "427fa8266f57c863fb0ec6d5e5a44b0a"
 }
