+++ morphir/runtime/jvm/src/org/finos/morphir/runtime/EvaluationLibraryPlatformSpecific.scala
@@ -18,30 +18,12 @@ import org.finos.morphir.runtime.quick.{EvaluatorQuick, Store}
 import org.finos.morphir.runtime.MorphirRuntimeError.*
 
 trait EvaluationLibraryPlatformSpecific {
-  def apply(fileName: String, prefix: Option[String] = None): EvaluationLibrary = {
-    val text = Source
-      .fromFile(fileName)
-      .getLines()
-      .mkString("\n")
-    val morphirIRFile = text.fromJson[MorphirIRFile]
-    val distribution = morphirIRFile
-      .getOrElse(throw new Exception(morphirIRFile.toString))
-      .distribution
-    EvaluationLibrary(MorphirRuntime.quick(distribution), prefix)
-  }
 
-  def apply(fileName: String, prefix: String): EvaluationLibrary = apply(fileName, Some(prefix))
-
-  def loadDistribution(fileName: String): Distribution = {
-    val text = Source
-      .fromFile(fileName)
-      .getLines()
-      .mkString("\n")
-    val morphirIRFile = text.fromJson[MorphirIRFile]
-    morphirIRFile
-      .getOrElse(throw new Exception(s"Failed to load $fileName as distribution"))
-      .distribution
-  }
+  def loadDistribution(fileName: String): Distribution =
+    Unsafe.unsafe {
+      implicit unsafe =>
+        Runtime.default.unsafe.run(loadDistributionFromFileZIO(fileName)).getOrThrowFiberFailure()
+    }
 
   def loadDistributionFromFileZIO(fileName: String): Task[Distribution] =
     for {
